+++ core/shared/src/test/scala/fs2/StreamSpec.scala
@@ -17,155 +17,6 @@ class StreamSpec extends Fs2Spec {
 
   "Stream" - {
 
-    "groupAdjacentBy" in forAll { (s: Stream[Pure, Int], n0: PosInt) =>
-      val n = n0 % 20 + 1
-      val f = (i: Int) => i % n
-      val s1 = s.groupAdjacentBy(f)
-      val s2 = s.map(f).changes
-      assert(s1.map(_._2).toList.flatMap(_.toList) == s.toList)
-      assert(s1.map(_._1).toList == s2.toList)
-      assert(
-        s1.map { case (k, vs) => vs.toVector.forall(f(_) == k) }.toList ==
-          s2.map(_ => true).toList
-      )
-    }
-
-    "groupAdjacentByLimit" - {
-      "limiting" in forAll { (s: Stream[Pure, Int], n0: PosInt) =>
-        val n = n0 % 20 + 1
-        val s1 = s.groupAdjacentByLimit(n)(_ => true)
-        assert(s1.map(_._2).toList.map(_.toList) == s.toList.grouped(n).toList)
-      }
-    }
-
-    "groupWithin" - {
-      "should never lose any elements" in forAll {
-        (s0: Stream[Pure, PosInt], d0: PosInt, maxGroupSize0: PosZInt) =>
-          val maxGroupSize = maxGroupSize0 % 20 + 1
-          val d = (d0 % 50).millis
-          val s = s0.map(_ % 500)
-          val sList = s.toList
-          s.covary[IO]
-            .evalTap(shortDuration => IO.sleep(shortDuration.micros))
-            .groupWithin(maxGroupSize, d)
-            .flatMap(s => Stream.emits(s.toList))
-            .compile
-            .toList
-            .asserting(it => assert(it == sList))
-      }
-
-      "should never emit empty groups" in forAll {
-        (s: Stream[Pure, PosInt], d0: PosInt, maxGroupSize0: PosZInt) =>
-          val maxGroupSize = maxGroupSize0 % 20 + 1
-          val d = (d0 % 50).millis
-          Stream(PosInt(1))
-            .append(s)
-            .map(_ % 500)
-            .covary[IO]
-            .evalTap(shortDuration => IO.sleep(shortDuration.micros))
-            .groupWithin(maxGroupSize, d)
-            .map(_.toList)
-            .compile
-            .toList
-            .asserting(it => assert(it.forall(_.nonEmpty)))
-      }
-
-      "should never have more elements than in its specified limit" in forAll {
-        (s: Stream[Pure, PosInt], d0: PosInt, maxGroupSize0: PosZInt) =>
-          val maxGroupSize = maxGroupSize0 % 20 + 1
-          val d = (d0 % 50).millis
-          s.map(_ % 500)
-            .evalTap(shortDuration => IO.sleep(shortDuration.micros))
-            .groupWithin(maxGroupSize, d)
-            .map(_.toList.size)
-            .compile
-            .toList
-            .asserting(it => assert(it.forall(_ <= maxGroupSize)))
-      }
-
-      "should return a finite stream back in a single chunk given a group size equal to the stream size and an absurdly high duration" in forAll {
-        (streamAsList0: List[Int]) =>
-          val streamAsList = 0 :: streamAsList0
-          Stream
-            .emits(streamAsList)
-            .covary[IO]
-            .groupWithin(streamAsList.size, (Int.MaxValue - 1L).nanoseconds)
-            .compile
-            .toList
-            .asserting(it => assert(it.head.toList == streamAsList))
-      }
-    }
-
-    "head" in forAll((s: Stream[Pure, Int]) => assert(s.head.toList == s.toList.take(1)))
-
-    "interleave" - {
-      "interleave left/right side infinite" in {
-        val ones = Stream.constant("1")
-        val s = Stream("A", "B", "C")
-        assert(ones.interleave(s).toList == List("1", "A", "1", "B", "1", "C"))
-        assert(s.interleave(ones).toList == List("A", "1", "B", "1", "C", "1"))
-      }
-
-      "interleave both side infinite" in {
-        val ones = Stream.constant("1")
-        val as = Stream.constant("A")
-        assert(ones.interleave(as).take(3).toList == List("1", "A", "1"))
-        assert(as.interleave(ones).take(3).toList == List("A", "1", "A"))
-      }
-
-      "interleaveAll left/right side infinite" in {
-        val ones = Stream.constant("1")
-        val s = Stream("A", "B", "C")
-        assert(
-          ones.interleaveAll(s).take(9).toList == List(
-            "1",
-            "A",
-            "1",
-            "B",
-            "1",
-            "C",
-            "1",
-            "1",
-            "1"
-          )
-        )
-        assert(
-          s.interleaveAll(ones).take(9).toList == List(
-            "A",
-            "1",
-            "B",
-            "1",
-            "C",
-            "1",
-            "1",
-            "1",
-            "1"
-          )
-        )
-      }
-
-      "interleaveAll both side infinite" in {
-        val ones = Stream.constant("1")
-        val as = Stream.constant("A")
-        assert(ones.interleaveAll(as).take(3).toList == List("1", "A", "1"))
-        assert(as.interleaveAll(ones).take(3).toList == List("A", "1", "A"))
-      }
-
-      // Uses a small scope to avoid using time to generate too large streams and not finishing
-      "interleave is equal to interleaveAll on infinite streams (by step-indexing)" in {
-        forAll(intsBetween(0, 100)) { (n: Int) =>
-          val ones = Stream.constant("1")
-          val as = Stream.constant("A")
-          assert(
-            ones.interleaveAll(as).take(n).toVector == ones
-              .interleave(as)
-              .take(n)
-              .toVector
-          )
-        }
-      }
-    }
-
     "interrupt" - {
       val interruptRepeatCount = if (isJVM) 25 else 1
